pyproject.toml
[tool.poetry]
name = "news-api"
version = "0.1.0"
description = ""
authors = ["Ajit Nath <ajit.nath@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.2"
aiohttp = "^3.8.1"
httpx = "^0.21.1"
SQLAlchemy = "^1.4.27"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.11b1"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
